+++ CHANGELOG.md
@@ -6,6 +6,7 @@ This changelog is inspired by [Keep a Changelog](https://keepachangelog.com/en/1
 ## 4.1.0 - 2023-07-22
 
 - changed: updated the list of parent activities (`PARENT_TYPE_ID`) to the new set Garmin is now using
+- changed: minimum Python version is now 3.8 (as Python 3.7 had its end-of-life on 2023-06-27)
 
 
 ## 4.0.0 - 2022-10-06